requirements.txt
altair==4.2.0
altgraph==0.17
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
astroid==2.4.2
asttokens==2.0.5
attrs==21.4.0
autopep8==1.5.5
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.9.3
black==22.1.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.0
cfgv==3.2.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
csvwriter==0.2.2
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.1
Django==4.0.2
docutils==0.16
entrypoints==0.4
executing==0.8.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
gTTS==2.2.2
h5py==3.2.1
identify==1.6.1
idna==2.10
imageai==2.1.5
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.1
jieba3k==0.35.1
Jinja2==2.11.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
newspaper3k==0.2.8
nltk==3.5
nodeenv==1.5.0
nose==1.3.7
notebook==6.4.8
numpy==1.20.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==20.9
pandas==1.2.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.1.1
platformdirs==2.5.0
pluggy==0.13.1
pre-commit==2.10.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.1
pure-eval==0.2.2
py==1.10.0
pyarrow==7.0.0
pycodestyle==2.6.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.8.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyjokes==0.6.0
pylint==2.6.2
pymongo==3.11.4
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==2.0.2
PyYAML==5.4.1
pyzmq==22.3.0
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
scikit-learn==1.0.2
scipy==1.6.2
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.2
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.5.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tox==3.22.0
tqdm==4.58.0
traitlets==5.1.1
translate==3.5.0
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.3
validators==0.18.2
virtualenv==20.4.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xlrd==2.0.1
xmltodict==0.12.0
